fixed soldier type in 2 q

+++ backend/military_unit/src/main/java/ccfit/nsu/ru/daryaevd/military_unit/dto/SoldierDto.java
@@ -19,7 +19,7 @@ public class SoldierDto {
     private String militaryCard;
     private Date dateOfIssueMilitaryCard;
     private Long masId;
-    private Long typeOfSoldier;
+    private Integer typeOfSoldier;
     private Long subdivisionId;
     private Boolean isCommander;
 }
